requirements/prod-requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
diff-match-patch==20230430
Django==5.0.7
django-admin-sortable2==2.2.1
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.1
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-mptt==0.16.0
django-unfold==0.38.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
firebase-admin==6.5.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googlemaps==4.10.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
import-export==0.3.1
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
multidict==6.1.0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyfcm==2.0.7
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-decouple==3.8
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.11.1
zope.interface==7.0.1
